src/requirements.txt
musdb
norbert
librosa
youtube-dl
museval
pydub
pytorch_lightning==0.9.0
soundfile
wandb
ffprobe
stempeg